Refactored to create single global scope object as per REDCap E.M. recommendation:
it is highly recommended that your JavaScript variables and functions not be placed in the global scope. Doing so could cause a conflict with other modules that are running at the same time that might have the same variable/function names. As an alternative, consider creating a function as an IIFE (Immediately Invoked Function Expression) or instead creating the variables/functions as properties of a single global scope object for the module, as seen below.